Allow multiple artifacts to exist (#102)
* Allow multiple artifacts to exist

I'm planning to publish the token in GitHub Pages soon. The only
realistic way to do that is to use actions/upload-pages-artifact
and then actions/deploy-pages -- the currently uploaded artifact
is not usable for this.

This commit makes the script not fail if there are multiple
artifacts, but still uses the same artifact as before. This makes it
possible to start publishing a new 'github-pages' artifact without
immediately breaking this script.

Signed-off-by: Jussi Kukkonen <jkukkonen@google.com>

* refactor oidc artifact id lookup

---------

Signed-off-by: Jussi Kukkonen <jkukkonen@google.com>