Add tests for verifying a C2SP-compliant checkpoint
The test for a valid checkpoint with a single signature is already
tested in the happy path. This adds positive tests for:

* A checkpoint with multiple signatures from the log, which could happen
  when a log has multiple signing algorithms, e.g. for PQC
* A checkpoint with a cosignature from a witness
* A checkpoint with two cosignatures from different witnesses
* A checkpoint with both multiple log signatures and a cosignature
* A checkpoint where the log signature is second and a witness signature
  is first, to check that there are no assumptions on ordering

This also adds negative tests for:

* Missing origin
* Missing size
* Missing root hash
* Missing log signature

I did not add positive tests for a log signed with ECDSA or RSA since
we'd need to generate a custom trusted root.

Signed-off-by: Hayden B <8418760+haydentherapper@users.noreply.github.com>