Manage python versions (#267)
* build: Improve virtual env management

Currently developer experience sucks since the project may or may not
run on a specific python version for various reasons. We don't really
want to start extensively testing on multiple versions but we can start
using a standard version

Use uv in the Makefile to make sure both virtual envs use a specified
Python version. This is done with a "bootstrap" virtual env that runs on
whatever python happens to be available.

Signed-off-by: Jussi Kukkonen <jkukkonen@google.com>

* Update requirements.txt

Signed-off-by: Jussi Kukkonen <jkukkonen@google.com>

---------

Signed-off-by: Jussi Kukkonen <jkukkonen@google.com>