Various small fixes (#294)
* action: Upload results even on failure

We'd rather see failures in the report than missing data.

Signed-off-by: Jussi Kukkonen <jkukkonen@google.com>

* report generation: Avoid checkmarks if no results available

Signed-off-by: Jussi Kukkonen <jkukkonen@google.com>

* report generation: Remove action version for now

Currently the action version is incorrect (git describe gives the
client tags, not conformance tags). We may want to redo this
with a version number from python sources but for now remove it.

Signed-off-by: Jussi Kukkonen <jkukkonen@google.com>

* Revert "Bump pytest from 8.4.2 to 9.0.0 (#290)"

This reverts commit 183d5af108cbd417b67ff20a493e774beadbaf20.

Somehow the new pytest manages to import things from the client
directory whatever action.py does ... revert upgrade for now.

Signed-off-by: Jussi Kukkonen <jkukkonen@google.com>

---------

Signed-off-by: Jussi Kukkonen <jkukkonen@google.com>